refact: move PerfApp to utils
